requirements.txt
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
dj-database-url==2.2.0
psycopg[binary]==3.2.3
whitenoise==6.8.2
gunicorn==23.0.0
celery==5.4.0
redis==5.2.1
django-ratelimit==4.1.0
requests==2.32.3
